Orangine/pyenv-requirements.txt
numpy==1.18.0
ptvsd==4.3.2
vectormath==0.2.2
jsonschema==3.0.2
argparse==1.4.0
mypy==0.910
ruamel.yaml==0.17.20